+++ cron/concerto.session.clear.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
-. /root/env.sh
-/usr/bin/php /app/concerto/bin/console concerto:sessions:clear --env=prod
+(
+  flock -n 9 || exit 0
+
+  . /root/env.sh
+  /usr/bin/php /app/concerto/bin/console concerto:sessions:clear --env=prod
+) 9>/data/lock/concerto.session.clear.lock