+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.init.R
@@ -22,11 +22,14 @@ concerto.init = function(dbConnectionParams, publicDir, platformUrl, appUrl, max
     RUNNER_PERSISTENT <<- 0
     RUNNER_SERIALIZED <<- 1
 
+    tempdir(T)
+
     concerto <<- list()
     concerto$cache <<- list(tests=list(), templates=list(), tables=list())
     concerto$cacheEnabled <<- T
     concerto$globals <<- list()
     concerto$templateParams <<- list()
+    concerto$globalTemplateParams <<- list()
     concerto$flow <<- list()
     concerto$flowIndex <<- 0
     concerto$bgWorkers <<- list()